(feature) Initial Commit - basic layout
This is the main shape of the Gem, including Spec setup and the dummy
rails app. Should be fairly simple to TDD and copy/pasta the old
spina-block functionality over.

This should make the implementation better, and allow us to validate it
did what it was supposed to, without really 'backfilling'.